Pulling out common device helper functions and generalizing some device macros.
