+++ src/chai/ArrayManager.hpp
@@ -7,122 +7,10 @@
 #ifndef CHAI_ArrayManager_HPP
 #define CHAI_ArrayManager_HPP
 
-#include "chai/config.hpp"
-#include "chai/ChaiMacros.hpp"
-#include "chai/ExecutionSpaces.hpp"
-#include "chai/PointerRecord.hpp"
-#include "chai/Types.hpp"
-
-#if defined(CHAI_ENABLE_RAJA_PLUGIN)
-#include "chai/pluginLinker.hpp"
-#endif
-
-#include <unordered_map>
-
-#include "umpire/Allocator.hpp"
-#include "umpire/util/MemoryMap.hpp"
-
-#if defined(CHAI_ENABLE_CUDA)
-#include <cuda_runtime_api.h>
-#endif
-#if defined(CHAI_ENABLE_HIP)
-#include "hip/hip_runtime_api.h"
-#endif
+#include "chai/ChaiManager.hpp"
 
 namespace chai
 {
-// CHAI_GPU_ERROR_CHECK macro
-#if defined(CHAI_ENABLE_CUDA) || defined(CHAI_ENABLE_HIP)
-
-#ifdef CHAI_ENABLE_GPU_ERROR_CHECKING
-
-#ifdef CHAI_ENABLE_CUDA
-inline void gpuErrorCheck(cudaError_t code, const char *file, int line, bool abort=true)
-{
-   if (code != cudaSuccess) {
-      fprintf(stderr, "[CHAI] GPU Error: %s %s %d\n", cudaGetErrorString(code), file, line);
-      if (abort) {
-         exit(code);
-      }
-   }
-}
-#elif defined(CHAI_ENABLE_HIP)
-inline void gpuErrorCheck(hipError_t code, const char *file, int line, bool abort=true)
-{
-   if (code != hipSuccess) {
-      fprintf(stderr, "[CHAI] GPU Error: %s %s %d\n", hipGetErrorString(code), file, line);
-      if (abort) {
-         exit(code);
-      }
-   }
-}
-#endif
-
-
-#define CHAI_GPU_ERROR_CHECK(code) { gpuErrorCheck((code), __FILE__, __LINE__); }
-#else // CHAI_ENABLE_GPU_ERROR_CHECKING
-#define CHAI_GPU_ERROR_CHECK(code) code
-#endif // CHAI_ENABLE_GPU_ERROR_CHECKING
-
-#endif
-
-// wrapper for hip/cuda synchronize
-inline void synchronize() {
-#if defined (CHAI_ENABLE_HIP) &&!defined(__HIP_DEVICE_COMPILE__)
-   CHAI_GPU_ERROR_CHECK(hipDeviceSynchronize());
-#elif defined (CHAI_ENABLE_CUDA) &&!defined(__CUDA_ARCH__)
-   CHAI_GPU_ERROR_CHECK(cudaDeviceSynchronize());
-#endif
-}
-
-#if defined(CHAI_GPUCC) || defined(CHAI_ENABLE_GPU_SIMULATION_MODE)
-
-// wrapper for hip/cuda free
-CHAI_HOST inline void gpuFree(void* buffer) {
-#if defined(CHAI_ENABLE_GPU_SIMULATION_MODE)
-   free(buffer);
-#elif defined (CHAI_ENABLE_HIP)
-   CHAI_GPU_ERROR_CHECK(hipFree(buffer));
-#elif defined (CHAI_ENABLE_CUDA)
-   CHAI_GPU_ERROR_CHECK(cudaFree(buffer));
-#endif
-}
-
-// wrapper for hip/cuda malloc
-CHAI_HOST inline void gpuMalloc(void** devPtr, size_t size) {
-#if defined(CHAI_ENABLE_GPU_SIMULATION_MODE)
-   *devPtr = (void*)malloc(size);
-#elif defined (CHAI_ENABLE_HIP)
-   CHAI_GPU_ERROR_CHECK(hipMalloc(devPtr, size));
-#elif defined (CHAI_ENABLE_CUDA)
-   CHAI_GPU_ERROR_CHECK(cudaMalloc(devPtr, size));
-#endif
-}
-
-// wrapper for hip/cuda managed malloc
-CHAI_HOST inline void gpuMallocManaged(void** devPtr, size_t size) {
-#if defined(CHAI_ENABLE_GPU_SIMULATION_MODE)
-   *devPtr = (void*)malloc(size);
-#elif defined (CHAI_ENABLE_HIP)
-   CHAI_GPU_ERROR_CHECK(hipMallocManaged(devPtr, size));
-#elif defined (CHAI_ENABLE_CUDA)
-   CHAI_GPU_ERROR_CHECK(cudaMallocManaged(devPtr, size));
-#endif
-}
-
-// wrapper for hip/cuda mem copy
-CHAI_HOST inline void  gpuMemcpy(void* dst, const void* src, size_t count, gpuMemcpyKind kind) {
-#if defined(CHAI_ENABLE_GPU_SIMULATION_MODE)
-   memcpy(dst, src, count);
-#elif defined (CHAI_ENABLE_HIP)
-   CHAI_GPU_ERROR_CHECK(hipMemcpy(dst, src, count, kind));
-#elif defined (CHAI_ENABLE_CUDA)
-   CHAI_GPU_ERROR_CHECK(cudaMemcpy(dst, src, count, kind));
-#endif
-}
-
-#endif //#if defined(CHAI_GPUCC)
-
 /*!
  * \brief Singleton that manages caching and movement of ManagedArray objects.
  *
