+++ src/chai/ChaiManager.hpp
@@ -1,0 +1,28 @@
+//////////////////////////////////////////////////////////////////////////////
+// Copyright (c) 2016-24, Lawrence Livermore National Security, LLC and CHAI
+// project contributors. See the CHAI LICENSE file for details.
+//
+// SPDX-License-Identifier: BSD-3-Clause
+//////////////////////////////////////////////////////////////////////////////
+#ifndef CHAI_ChaiManager_HPP
+#define CHAI_ChaiManager_HPP
+
+#include "chai/ChaiMacros.hpp"
+#include "chai/ExecutionSpaces.hpp"
+#include "chai/Types.hpp"
+
+#include "chai/PointerRecord.hpp"
+
+#if defined(CHAI_ENABLE_RAJA_PLUGIN)
+#include "chai/pluginLinker.hpp"
+#endif
+
+#include <unordered_map>
+
+#include "umpire/Allocator.hpp"
+#include "umpire/util/MemoryMap.hpp"
+
+
+#include "chai/util/DeviceHelpers.hpp"
+
+#endif // CHAI_ChaiManager_HPP
